+++ test/controllers/internal_controller_test.rb
@@ -13,4 +13,22 @@ class InternalControllerTest < ActionDispatch::IntegrationTest
     get dashboard_url
     assert_response :success
   end
+
+  test "should get reports" do
+    login(@viewer)
+    get reports_url
+    assert_response :success
+  end
+
+  test "should get data health" do
+    login(@viewer)
+    get data_health_url
+    assert_response :success
+  end
+
+  test "should get report card" do
+    login(@viewer)
+    get report_card_url
+    assert_response :success
+  end
 end